+++ src/components/InfosUsuario.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+export function InfosUsuario() {
+    const usuario = 'Marcus'
+    return(
+        <>
+            <img src="https://picsum.photos/400/400?a=1 " alt="" />
+            <h3>Feito por {usuario}</h3>
+        </>
+    )
+}
